requirements.txt
python-telegram-bot==20.7
aiosqlite==0.19.0
hypothesis==6.92.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
